+++ test/sparkplug_device_command_spec.js
@@ -101,7 +101,6 @@ describe('mqtt sparkplug device node - Commands', function () {
 		  });
 
 		client.on('message', function (topic, message) {
-			console.log(topic);
 			if (topic === "spBv1.0/My Devices/DBIRTH/Node-Red/NEW_NAME") {
 				var buffer = Buffer.from(message);
 				var payload = spPayload.decodePayload(buffer);
@@ -179,4 +178,52 @@ describe('mqtt sparkplug device node - Commands', function () {
 			}	
 		});
     });
+
+
+	it('should not birth until connect', function (done) {
+	
+	flow = simpleFlow;
+	flow[1].manualEoNBirth = true;
+	client = mqtt.connect(testBroker);
+
+	let n1;
+	let b1;
+	var waitOver = false;
+	client.on('connect', function () {
+		client.subscribe('#', function (err) {
+		  if (!err) {
+			helper.load(sparkplugNode, simpleFlow, function () {
+				
+
+				try {
+					n1 = helper.getNode("n1");
+					b1 = n1.brokerConn;
+
+					setTimeout(() => {
+						waitOver = true;
+						n1.receive({
+							"command" : {
+								"EoN" : {
+									"connect" : true
+								}
+							}
+						})	
+					}, 500);
+				}catch (e) {
+					done(e);
+				}
+			});
+		  }
+		})
+	  });
+
+	
+
+	client.on('message', function (topic, message) {
+		if (topic === "spBv1.0/My Devices/NBIRTH/Node-Red") {
+			waitOver.should.be.true();
+			done();
+		}
+	});
+});
 });